+++ packages/excalidraw/package.json
@@ -2,7 +2,7 @@
   "name": "@excalidraw/excalidraw",
   "version": "0.17.1",
   "main": "main.js",
-  "types": "types/packages/excalidraw/index.d.ts",
+  "types": "types/excalidraw/index.d.ts",
   "files": [
     "dist/*",
     "types/*"
@@ -77,9 +77,6 @@
     "tunnel-rat": "0.1.2"
   },
   "devDependencies": {
-    "@types/pako": "1.0.3",
-    "@types/pica": "5.1.3",
-    "@types/resize-observer-browser": "0.1.7",
     "@babel/core": "7.18.9",
     "@babel/plugin-transform-arrow-functions": "7.18.6",
     "@babel/plugin-transform-async-to-generator": "7.18.6",
@@ -89,6 +86,9 @@
     "@babel/preset-react": "7.18.6",
     "@babel/preset-typescript": "7.18.6",
     "@size-limit/preset-big-lib": "9.0.0",
+    "@types/pako": "1.0.3",
+    "@types/pica": "5.1.3",
+    "@types/resize-observer-browser": "0.1.7",
     "autoprefixer": "10.4.7",
     "babel-loader": "8.2.5",
     "babel-plugin-transform-class-properties": "6.24.1",
