+++ packages/utils/withinBounds.test.ts
@@ -1,5 +1,5 @@
-import { Bounds } from "./excalidraw/element/bounds";
-import { API } from "./excalidraw/tests/helpers/api";
+import { Bounds } from "../excalidraw/element/bounds";
+import { API } from "../excalidraw/tests/helpers/api";
 import {
   elementPartiallyOverlapsWithOrContainsBBox,
   elementsOverlappingBBox,
